+++ pom.xml
@@ -47,7 +47,7 @@
                             <addClasspath>true</addClasspath>
                             <classpathPrefix>libs/</classpathPrefix>
                             <mainClass>
-                                quiztastic.server.JeopardyServer
+                                quiztastic.entries.JeopardyServer
                             </mainClass>
                         </manifest>
                     </archive>
